vcms/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2.7
django-sequences==2.8
djangorestframework==3.14.0
fontawesomefree==6.5.1
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.26.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
idna==3.6
mysqlclient==2.2.0
oauthlib==3.2.2
Pillow==10.1.0
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyparsing==3.1.1
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.4
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.1.0
